rename storage browser to web storage which is easier to use and remember
